refactor: move logic to hooks file
